Added header and import comments. selectPoi() in map.dart updates MapControllerService().
