+++ lib/models/poi.dart
@@ -1,3 +1,6 @@
+// Copyright (c) 2026, Owen Danke
+
+// pub.dev
 import 'package:latlong2/latlong.dart';
 
 class PointOfInterest {
