+++ lib/pages/explore/tree_list.dart
@@ -1,3 +1,5 @@
+// Copyright (c) 2026, Owen Danke
+
 // Flutter
 import 'package:flutter/material.dart';
