+++ lib/pages/home/visitor_info.dart
@@ -1,5 +1,10 @@
+// Copyright (c) 2026, Owen Danke
+
+// Flutter
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+
+// httapp
 import 'package:httapp/services/text_theme_service.dart';
 
 class VisitorInfoPage extends StatelessWidget {
