+++ lib/routes/map_routes.dart
@@ -1,8 +1,11 @@
-import 'package:flutter/material.dart';
-import 'app_routes.dart';
-import 'package:httapp/main.dart';
+// Copyright (c) 2026, Owen Danke
 
-// Import page files
+// Flutter
+import 'package:flutter/material.dart';
+
+// httapp
+import 'package:httapp/routes/app_routes.dart';
+import 'package:httapp/main.dart';
 import 'package:httapp/pages/map/map.dart';
 import 'package:httapp/ui/template_tree_page.dart';
