+++ lib/routes/settings_routes.dart
@@ -1,7 +1,10 @@
-import 'package:flutter/material.dart';
-import 'app_routes.dart';
+// Copyright (c) 2026, Owen Danke
 
-// Import page files
+// Flutter
+import 'package:flutter/material.dart';
+
+// httapp
+import 'package:httapp/routes/app_routes.dart';
 import 'package:httapp/pages/settings/settings.dart';
 import 'package:httapp/pages/settings/advanced_settings.dart';
 import 'package:httapp/pages/settings/customization_settings.dart';
